requirements.txt
pydantic>=2.0.2
sanic==23.6.0
sanic-ext==23.6.0
loguru>=0.7.0
vyper-config>=1.1.1
click==8.1.6
pymongo>=4.4.1
motor>=3.2.0
email-validator
sanic-validation
sanic-jwt
bcrypt>=4.0.1
pyjwt>=2.8.0
pyyaml>=6.0
shortuuid>=1.0.11
kubernetes>=27.2.0
mongoquery>=1.4.2
jsonpath-ng>=1.5.3
httpx>=0.23.3
urllib3~=2.2.2
google-auth~=2.30.0
odmantic~=1.0.2
casbin~=1.36.3
# casbin_pymongo_adapter~=1.1.0
aiohttp~=3.9.5
aiohttp_retry~=2.8.3
pyotp~=2.9.0
